+++ src/Program.fs
@@ -1,5 +1,6 @@
-﻿// Learn more about F# at http://fsharp.org
+// Learn more about F# at http://fsharp.org
 
+open System
 open MF.ConsoleApplication
 open MF.ImageManager
 
@@ -17,6 +18,40 @@ let main argv =
             Initialize = None
             Interact = None
             Execute = Command.PrepareCommand.execute
+        }
+
+        command "about" {
+            Description = "Displays information about the current project."
+            Help = None
+            Arguments = []
+            Options = []
+            Initialize = None
+            Interact = None
+            Execute = fun (_input, output) ->
+                let ``---`` = [ "------------------"; "----------------------------------------------------------------------------------------------" ]
+
+                output.Table [ AssemblyVersionInformation.AssemblyProduct ] [
+                    [ "Description"; AssemblyVersionInformation.AssemblyDescription ]
+                    [ "Version"; AssemblyVersionInformation.AssemblyVersion ]
+
+                    ``---``
+                    [ "Environment" ]
+                    ``---``
+                    [ ".NET Core"; Environment.Version |> sprintf "%A" ]
+                    [ "Command Line"; Environment.CommandLine ]
+                    [ "Current Directory"; Environment.CurrentDirectory ]
+                    [ "Machine Name"; Environment.MachineName ]
+                    [ "OS Version"; Environment.OSVersion |> sprintf "%A" ]
+                    [ "Processor Count"; Environment.ProcessorCount |> sprintf "%A" ]
+
+                    ``---``
+                    [ "Git" ]
+                    ``---``
+                    [ "Branch"; AssemblyVersionInformation.AssemblyMetadata_gitbranch ]
+                    [ "Commit"; AssemblyVersionInformation.AssemblyMetadata_gitcommit ]
+                ]
+
+                ExitCode.Success
         }
     }
     |> run argv