+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - Allow to exclude images from a file
 - Allow to dry-run the preparation
 - Allow to create a target subdirs by image created year/month
+- Allow config file